refactor(utils): group storage modules into package
