+++ utils/storage/__init__.py
@@ -1,0 +1,7 @@
+"""Storage layout, policy, migration, and startup coordination.
+
+Implementations live in this package.  The historical ``utils.storage_*``
+modules remain as compatibility aliases so existing imports and monkeypatch
+targets keep referring to the same module objects.
+"""
+
